requirements.txt
pyautogui
pyscreeze
pillow
aircv
numpy
opencv-python
